Changed provider structure and fixed imports/exports
*Imports/Exports*
The exported files in `lib/public_api.ts` could not beresolved, because it exported the folders, not the `index.ts` files (see https://github.com/ng-packagr/ng-packagr/issues/355#issuecomment-353002678)

*Provider Structure*
Changed `PrismicClientModule` and its service from a `forRoot()`-Approach to a custom `InjectionToken` for the config object. This fixes several issues with the Angular Compiler when importing/providing the module.